requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
fpdf==1.7.2
idna==2.10
-e git+https://github.com/Burbon13/Personal-Silly-CLI.git@28a230718b9d60dfea74a267e7deb0c466a86a1e#egg=my_silly_cli_tool
Pillow==8.4.0
pip-autoremove==0.9.1
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
six==1.15.0
tinydb==4.2.0
twilio==6.45.4
urllib3==1.25.10
validate-email==1.3
